training/pyproject.toml
[project]
name = "hw3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "ipython>=9.12.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "tokenizers>=0.22.2",
    # "torch>=2.11.0", # for gpu configs, might need to comment this one out
    "torch>=2.11.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.11",
]

[project.scripts]
train = "train:main"
fine-tune = "fine_tuning.fine_tune:main"
controlled-prompts = "fine_tuning.controlled_prompts:main"

# TODO: build_instruction_dataset + fill_instruction_prompts

[tool.setuptools.packages.find]
where = ["."]
